Move `disk`, `ball`, `ellipse` back to ski and add migration notes
